Refactor decorator tests into multiple files
