+++ tests/test_decorator_auth_required.py
@@ -1,0 +1,109 @@
+from openapi_spec_validator import validate_spec
+
+from apiflask import APIBlueprint
+from apiflask import auth_required
+from apiflask.security import HTTPBasicAuth
+from apiflask.security import HTTPTokenAuth
+
+
+def test_auth_required(app, client):
+    auth = HTTPBasicAuth()
+
+    @auth.verify_password
+    def verify_password(username, password):
+        if username == 'foo' and password == 'bar':
+            return {'user': 'foo'}
+        elif username == 'bar' and password == 'foo':
+            return {'user': 'bar'}
+        elif username == 'baz' and password == 'baz':
+            return {'user': 'baz'}
+
+    @auth.get_user_roles
+    def get_roles(user):
+        if user['user'] == 'bar':
+            return 'admin'
+        elif user['user'] == 'baz':
+            return 'moderator'
+        return 'normal'
+
+    @app.route('/foo')
+    @auth_required(auth)
+    def foo():
+        return auth.current_user
+
+    @app.route('/bar')
+    @auth_required(auth, role='admin')
+    def bar():
+        return auth.current_user
+
+    @app.route('/baz')
+    @auth_required(auth, roles=['admin', 'moderator'])
+    def baz():
+        return auth.current_user
+
+    rv = client.get('/foo')
+    assert rv.status_code == 401
+
+    rv = client.get('/foo', headers={'Authorization': 'Basic Zm9vOmJhcg=='})
+    assert rv.status_code == 200
+    assert rv.json == {'user': 'foo'}
+
+    rv = client.get('/bar', headers={'Authorization': 'Basic Zm9vOmJhcg=='})
+    assert rv.status_code == 403
+
+    rv = client.get('/foo', headers={'Authorization': 'Basic YmFyOmZvbw=='})
+    assert rv.status_code == 200
+    assert rv.json == {'user': 'bar'}
+
+    rv = client.get('/bar', headers={'Authorization': 'Basic YmFyOmZvbw=='})
+    assert rv.status_code == 200
+    assert rv.json == {'user': 'bar'}
+
+    rv = client.get('/baz', headers={'Authorization': 'Basic Zm9vOmJhcg=='})
+    assert rv.status_code == 403
+
+    rv = client.get('/baz', headers={'Authorization': 'Basic YmFyOmZvbw=='})
+    assert rv.status_code == 200
+    assert rv.json == {'user': 'bar'}
+
+    rv = client.get('/baz', headers={'Authorization': 'Basic YmF6OmJheg=='})
+    assert rv.status_code == 200
+    assert rv.json == {'user': 'baz'}
+
+
+def test_auth_required_at_blueprint_before_request(app, client):
+    bp = APIBlueprint('test', __name__)
+
+    auth = HTTPTokenAuth()
+
+    @bp.before_request
+    @auth_required(auth)
+    def before():
+        pass
+
+    @bp.get('/foo')
+    def foo():
+        pass
+
+    @bp.get('/bar')
+    def bar():
+        pass
+
+    app.register_blueprint(bp)
+
+    rv = client.get('/foo')
+    assert rv.status_code == 401
+    rv = client.get('/bar')
+    assert rv.status_code == 401
+
+    rv = client.get('/openapi.json')
+    assert rv.status_code == 200
+    validate_spec(rv.json)
+    assert 'BearerAuth' in rv.json['components']['securitySchemes']
+    assert rv.json['components']['securitySchemes']['BearerAuth'] == {
+        'scheme': 'Bearer',
+        'type': 'http'
+    }
+
+    assert 'BearerAuth' in rv.json['paths']['/foo']['get']['security'][0]
+    assert 'BearerAuth' in rv.json['paths']['/bar']['get']['security'][0]
